Clean up Google App-tracking code
To implement the Google crankd Application-tracking code, we're
going to need to have sqlite3 installed...the problem is that the
Rubygem is problematic on 10.5 and 10.6. To remedy this, we'll use
Macports to install it.  Unfortunately, Macports isn't yet INSTALLED
on the machines, so we'll need to install the developer tools and then
install Macports. That's what this series of commits will do.